Move useEffect to useGetIngreduentIds
